ENH Use Integer throughout

+++ waldo/locate/models.py
@@ -21,14 +21,14 @@
 #  THE SOFTWARE.
 
 from __future__ import division
-from sqlalchemy import Column, Boolean, String, Integer, BigInteger, ForeignKey
+from sqlalchemy import Column, Boolean, String, Integer, ForeignKey
 from sqlalchemy.orm import relation
 from waldo.backend import Base
 
 class Isoform(Base):
     __tablename__ = 'locate_isoforms'
     id = Column(Integer, primary_key=True)
-    locate_id = Column(BigInteger, ForeignKey('locate_entries.id'), index=True)
+    locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
     isoform_class = Column(String(20))
     isoform_name = Column(String(30))
 
@@ -39,7 +39,7 @@ class Isoform(Base):
 class Image(Base):
     __tablename__ = 'locate_images'
     id = Column(Integer, primary_key=True)
-    locate_id = Column(BigInteger, ForeignKey('locate_entries.id'), index=True)
+    locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
     filename = Column(String(50))
     is_coloc = Column(Boolean)
     celltype = Column(String(100))
@@ -67,8 +67,8 @@ class Image(Base):
 class LocatePrediction(Base):
     __tablename__ = 'locate_predictions'
     id = Column(Integer, primary_key=True)
-    locate_id = Column(BigInteger, ForeignKey('locate_entries.id'), index=True)
-    source_id = Column(BigInteger)
+    locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
+    source_id = Column(Integer)
     method = Column(String(30))
     location = Column(String(50))
     go_id = Column(String(50))
@@ -82,11 +82,11 @@ class LocatePrediction(Base):
 class Literature(Base):
     __tablename__ = 'locate_literature'
     id = Column(Integer, primary_key=True)
-    locate_id = Column(BigInteger, ForeignKey('locate_entries.id'), index=True)
+    locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
     authors = Column(String(200))
     title = Column(String(200))
     citation = Column(String(100))
-    source_id = Column(BigInteger)
+    source_id = Column(Integer)
     source_name = Column(String(50))
     accn = Column(String(20))
 
@@ -119,10 +119,10 @@ class Literature(Base):
 
 class LocateAnnotation(Base):
     __tablename__ = 'locate_annotations'
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
     evidence = Column(String(64))
-    source_id = Column(BigInteger)
+    source_id = Column(Integer)
     source_name = Column(String(50))
     accn = Column(String(20))
     go_id = Column(String(32))
@@ -139,8 +139,8 @@ class LocateAnnotation(Base):
 class ExternalReference(Base):
     __tablename__ = 'locate_externalreferences'
     xref_id = Column(Integer, primary_key=True)
-    locate_id = Column(BigInteger, ForeignKey('locate_entries.id'), index=True)
-    source_id = Column(BigInteger)
+    locate_id = Column(Integer, ForeignKey('locate_entries.id'), index=True)
+    source_id = Column(Integer)
     source_name = Column(String(50))
     accn = Column(String(20))
 
@@ -151,7 +151,7 @@ class ExternalReference(Base):
 
 class LocateEntry(Base):
     __tablename__ = 'locate_entries'
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     organism = Column(String(30), nullable=True)
     name = Column(String(50))
     isoforms = relation(Isoform)
